fixed the template to use cfn-init on the head node.
the bootstrap process downloads the cookbook from the s3 location and
installs it into /var/chef/cookbook

then calls cfn-init -s genome-alignment -r head to configure chef

then calls chef-solo to run the recipes.  currently the only recipe is
the ec2-commandline-tools, which is really a template for the other
software.

updated the makefile to pass in the AWS Access key and AWS Secret key.

% make create should work now along with
% make destroy to bring down the stack.